Convert "Library" pages to typescript
