+++ src/pages/library/books/SingleBook/index.tsx
@@ -1,84 +1,38 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { FormattedMessage } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
-import styled from 'styled-components';
 import { FiDownload as DownloadIcon } from 'react-icons/fi';
-
+import { getSingleItemRequest } from '../../../../actions/library';
 import {
-  Paper,
-  Container,
-  Button,
   LoadContent,
-  Dropdown,
   DropdownListContainer,
   DropdownListItem,
-} from '../../components/ui';
+} from '../../../../components/ui';
+import { Paper, Container, Button, Dropdown } from 'snake-ui';
 
-import defaultThumbnail from '../../assets/default-book-cover.jpg';
-import Categories from '../../components/categories/ShowCategories';
+import defaultThumbnail from '../../../../assets/default-book-cover.jpg';
+import Categories from '../../../../components/categories/ShowCategories';
 
-import { BookDisplayContainer, BookDisplayGrid } from './styles.css';
-
-import { getSingleItemRequest as getSingleItem } from '../../actions/library';
-
-const Title = styled.h2`
-  font-weight: bold;
-  color: ${(props) => props.theme.palette.text.contrast};
-  font-size: 2.125rem;
-  margin-bottom: 5px;
-`;
-
-const Author = styled.h3`
-  font-weight: lighter;
-  color: ${(props) => props.theme.palette.text.contrast};
-  font-size: 1.25rem;
-  margin-bottom: 34px;
-`;
-
-const BookCover = styled.img`
-  width: 100%;
-  @media only screen and (max-width: 768px) {
-    height: 240px;
-    object-fit: cover;
-  }
-`;
-
-const Banner = styled.div`
-  background: url(${(props) => (props.cover ? props.cover : defaultThumbnail)})
-    rgba(0, 0, 0, 0.5);
-  background-size: cover;
-  background-position: center;
-  background-attachment: fixed;
-  position: relative;
-  width: 100%;
-  height: 230px;
-
-  &:after {
-    height: 230px;
-    width: 100%;
-
-    background: linear-gradient(
-      to bottom,
-      rgba(0, 0, 0, 0.4) 0%,
-      rgba(0, 0, 0, 1) 100%
-    );
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 0;
-    opacity: 0.7;
-  }
-`;
+import {
+  BookDisplayContainer,
+  BookDisplayGrid,
+  Title,
+  Banner,
+  Author,
+  BookCover,
+} from './styles';
 
 function SingleBook() {
-  const { id } = useParams();
+  const { id }: { id: string } = useParams();
   const dispatch = useDispatch();
-  React.useEffect(() => {
-    dispatch(getSingleItem({ itemId: id }));
+  const [downloadAnchor, setDownloadAnchor] = useState(null);
+
+  useEffect(() => {
+    dispatch(getSingleItemRequest({ itemId: id }));
   }, [dispatch, id]);
 
-  const { singleItem, loading } = useSelector((state) => state.library);
+  const { singleItem, loading } = useSelector((state: any) => state.library);
 
   return (
     <LoadContent loading={loading}>
@@ -102,17 +56,21 @@ function SingleBook() {
                   }
                 />
                 <div style={{ padding: 10 }}>
+                  <Button
+                    color="secondary"
+                    onClick={(e: any) => setDownloadAnchor(e.currentTarget)}
+                  >
+                    <FormattedMessage id="common.download" />
+                  </Button>
                   <Dropdown
-                    toggle={
-                      <Button fullwidth color="secondary">
-                        <FormattedMessage id="common.download" />
-                      </Button>
-                    }
                     placement="top"
+                    anchorEl={downloadAnchor}
+                    onClose={() => setDownloadAnchor(null)}
+                    open={Boolean(downloadAnchor)}
                   >
                     <DropdownListContainer>
                       {singleItem.files &&
-                        singleItem.files.map((file) => (
+                        singleItem.files.map((file: any) => (
                           <DropdownListItem
                             icon={<DownloadIcon />}
                             key={file.originalname}
@@ -126,7 +84,6 @@ function SingleBook() {
                   </Dropdown>
                 </div>
               </Paper>
-              {/* <InvitedBy user={user} /> */}
             </div>
             <div>
               <div style={{ marginBottom: 8 }}>
@@ -136,7 +93,6 @@ function SingleBook() {
               <Title>{singleItem && singleItem.title}</Title>
               <Author>{singleItem && singleItem.author}</Author>
               <div>{singleItem && singleItem.content}</div>
-              <div my={2}>{/* <Ratings item={book.item} /> */}</div>
             </div>
           </BookDisplayGrid>
         </Container>
