+++ src/pages/library/index.tsx
@@ -9,36 +9,35 @@ import LibraryCard from '../../components/library/LibraryCard';
 import { loadCategoriesRequest } from '../../actions/categories';
 import { getItemsRequest } from '../../actions/library';
 
+import { MenuItem, LoadContent } from '../../components/ui';
 import {
   Container,
   Hero,
   Card,
   CardHeader,
-  MenuItem,
   Tab,
   Tabs,
-  LoadContent,
   Paper,
-} from '../../components/ui';
+  Grid,
+} from 'snake-ui';
 
 import {
   LibraryContainer,
   LibrarySidebarContainer,
   LibraryContentContainer,
   LibrarySidebarMenu,
-  LibraryCardGrid,
-} from './styles.css';
+} from './styles';
 
 const Library = () => {
   // eslint-disable-next-line no-unused-vars
   const [type, setType] = useState('');
   const [selectedCategory, setSelectedCategory] = useState('');
   const { items: categories, loading: loadingCategories } = useSelector(
-    (state) => state.categories
+    (state: any) => state.categories
   );
-  const { items, loading } = useSelector((state) => state.library);
+  const { items, loading } = useSelector((state: any) => state.library);
   const dispatch = useDispatch();
-  const { type: typeParam } = useParams();
+  const { type: typeParam }: { type: string } = useParams();
 
   useEffect(() => {
     dispatch(loadCategoriesRequest());
@@ -49,7 +48,7 @@ const Library = () => {
   }, [typeParam]);
 
   useEffect(() => {
-    const types = {
+    const types: { [key: string]: string } = {
       articles: 'article',
       books: 'book',
       videos: 'video',
@@ -64,6 +63,7 @@ const Library = () => {
       <Hero
         title={<FormattedMessage id="common.library" />}
         description={<FormattedMessage id="home.features.0" />}
+        actions
       />
 
       <LibraryContainer>
@@ -81,7 +81,7 @@ const Library = () => {
                   current={selectedCategory === ''}
                   onClick={() => setSelectedCategory('')}
                 />
-                {categories.map((category) => (
+                {categories.map((category: any) => (
                   <MenuItem
                     key={category.name}
                     label={category.name}
@@ -96,22 +96,22 @@ const Library = () => {
         <LibraryContentContainer>
           <Paper>
             <Tabs style={{ height: 48, padding: '0px 8px' }}>
-              <Tab
+              <Tab // @ts-ignore
                 label={<FormattedMessage id="common.all" />}
                 current={typeParam === undefined}
                 link="/library"
               />
-              <Tab
+              <Tab // @ts-ignore
                 label={<FormattedMessage id="library.articles" />}
                 current={typeParam === 'articles'}
                 link="/library/articles"
               />
-              <Tab
+              <Tab // @ts-ignore
                 label={<FormattedMessage id="library.books" />}
                 current={typeParam === 'books'}
                 link="/library/books"
               />
-              <Tab
+              <Tab // @ts-ignore
                 label={<FormattedMessage id="library.videos" />}
                 current={typeParam === 'videos'}
                 link="/library/videos"
@@ -125,13 +125,13 @@ const Library = () => {
                   <FormattedMessage id="library.noneFound" />
                 </Paper>
               ) : (
-                <LibraryCardGrid>
-                  {items.map((item) => (
-                    <div xs={12} key={() => generate()}>
+                <Grid container spacing={2}>
+                  {items.map((item: any) => (
+                    <Grid xs={12} key={generate()}>
                       <LibraryCard item={item} />
-                    </div>
+                    </Grid>
                   ))}
-                </LibraryCardGrid>
+                </Grid>
               )}
             </LoadContent>
           </div>
